Test that we can run unit tests
Before we go too far we want to make sure our test framework is working.
To do this we can run `pytest` in our virtual environment or setup out
IDE.

In my case I am using vscode. In order to setup testing in vscode I go
to the Command Palette (Command+Shift+P on OS X), and choose "Python:
Configure Tests". From there I choose pytest and choose the directory in
which my tests exist. In this case that is the test directory.

Once I do all of that I can run tests inside of vscode using the Run
tests button at the bottom, right clicking the test file and choose run
test file, or clicking on the beaker icon on the left and clicking the
Run All Tests button.